Move sha256 checksum check to audit
